refactor: use `Predicate`
